// src/app/Interfaces/View/all-columns-customer-compitions.ts
export interface AllColumnsCustomerCompitions {
    competitionCompanyId:number;
    competitionCompanyName:string;
    customerName:string;
    customerID: number;
    territoryName: string;
    sectorName: string;
    governorateName: string;
    regionName: string;
    company : string;
    adress : string;
    territoryID: number;
    sectorID: number;
    governorateID: number;
    regionID: number;
    total:number;
}